[minSizeCC / maxSizeCC] Use tabs for indentation, for consistency
