+++ src/main/java/org/chocosolver/graphsolver/variables/GraphVar.java
@@ -43,358 +43,358 @@ import java.lang.reflect.Field;
 
 public abstract class GraphVar<E extends IGraph> extends AbstractVariable implements Variable {
 
-    public static final int GRAPH = 1 << 7; // beware, this relies on choco-solver
+	public static final int GRAPH = 1 << 7; // beware, this relies on choco-solver
 
-    //////////////////////////////// GRAPH PART /////////////////////////////////////////
-    //***********************************************************************************
-    // VARIABLES
-    //***********************************************************************************
+	//////////////////////////////// GRAPH PART /////////////////////////////////////////
+	//***********************************************************************************
+	// VARIABLES
+	//***********************************************************************************
 
-    protected E UB, LB;
-    protected GraphDelta delta;
-    protected int n;
-    ///////////// Attributes related to Variable ////////////
-    protected boolean reactOnModification;
+	protected E UB, LB;
+	protected GraphDelta delta;
+	protected int n;
+	///////////// Attributes related to Variable ////////////
+	protected boolean reactOnModification;
 
-    //***********************************************************************************
-    // CONSTRUCTORS
-    //***********************************************************************************
+	//***********************************************************************************
+	// CONSTRUCTORS
+	//***********************************************************************************
 
-    /**
-     * Creates a graph variable
-     *
-     * @param solver
-     */
-    protected GraphVar(String name, Model solver, E LB, E UB) {
-        super(name, solver);
-        this.LB = LB;
-        this.UB = UB;
-        this.n = UB.getNbMaxNodes();
-        assert n == LB.getNbMaxNodes();
-        Field f = null; //NoSuchFieldException
-        try {
-            AbstractVariable me = this;
-            f = me.getClass().getSuperclass().getSuperclass().getDeclaredField("scheduler");
-            f.setAccessible(true);
-            f.set(me, new GraphEvtScheduler());
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            e.printStackTrace();
-        }
+	/**
+	 * Creates a graph variable
+	 *
+	 * @param solver
+	 */
+	protected GraphVar(String name, Model solver, E LB, E UB) {
+		super(name, solver);
+		this.LB = LB;
+		this.UB = UB;
+		this.n = UB.getNbMaxNodes();
+		assert n == LB.getNbMaxNodes();
+		Field f = null; //NoSuchFieldException
+		try {
+			AbstractVariable me = this;
+			f = me.getClass().getSuperclass().getSuperclass().getDeclaredField("scheduler");
+			f.setAccessible(true);
+			f.set(me, new GraphEvtScheduler());
+		} catch (NoSuchFieldException | IllegalAccessException e) {
+			e.printStackTrace();
+		}
 	}
 
-    //***********************************************************************************
-    // METHODS
-    //***********************************************************************************
+	//***********************************************************************************
+	// METHODS
+	//***********************************************************************************
 
-    @Override
-    public boolean isInstantiated() {
-        if (getPotentialNodes().size() != getMandatoryNodes().size()) {
-            return false;
-        }
-        ISet suc;
-        for (int i :getUB().getNodes()) {
-            suc = UB.getSuccOrNeighOf(i);
-            if (suc.size() != getLB().getSuccOrNeighOf(i).size()) {
-                return false;
-            }
-        }
-        return true;
-    }
+	@Override
+	public boolean isInstantiated() {
+		if (getPotentialNodes().size() != getMandatoryNodes().size()) {
+			return false;
+		}
+		ISet suc;
+		for (int i :getUB().getNodes()) {
+			suc = UB.getSuccOrNeighOf(i);
+			if (suc.size() != getLB().getSuccOrNeighOf(i).size()) {
+				return false;
+			}
+		}
+		return true;
+	}
 
-    /**
-     * Remove node x from the domain
-     * Removes x from the upper bound graph
-     *
-     * @param x     node's index
-     * @param cause algorithm which is related to the removal
-     * @return true iff the removal has an effect
-     */
-    public boolean removeNode(int x, ICause cause) throws ContradictionException {
-        assert cause != null;
-        assert (x >= 0 && x < n);
-        if (LB.getNodes().contains(x)) {
-            this.contradiction(cause, "remove mandatory node");
-            return true;
-        } else if (!UB.getNodes().contains(x)) {
-            return false;
-        }
-        ISet nei = UB.getSuccOrNeighOf(x);
-        for (int i : nei) {
-            removeArc(x, i, cause);
-        }
-        nei = UB.getPredOrNeighOf(x);
-        for (int i : nei) {
-            removeArc(i, x, cause);
-        }
-        if (UB.removeNode(x)) {
-            if (reactOnModification) {
-                delta.add(x, GraphDelta.NR, cause);
-            }
-            GraphEventType e = GraphEventType.REMOVE_NODE;
-            notifyPropagators(e, cause);
-            return true;
-        }
-        return false;
-    }
+	/**
+	 * Remove node x from the domain
+	 * Removes x from the upper bound graph
+	 *
+	 * @param x     node's index
+	 * @param cause algorithm which is related to the removal
+	 * @return true iff the removal has an effect
+	 */
+	public boolean removeNode(int x, ICause cause) throws ContradictionException {
+		assert cause != null;
+		assert (x >= 0 && x < n);
+		if (LB.getNodes().contains(x)) {
+			this.contradiction(cause, "remove mandatory node");
+			return true;
+		} else if (!UB.getNodes().contains(x)) {
+			return false;
+		}
+		ISet nei = UB.getSuccOrNeighOf(x);
+		for (int i : nei) {
+			removeArc(x, i, cause);
+		}
+		nei = UB.getPredOrNeighOf(x);
+		for (int i : nei) {
+			removeArc(i, x, cause);
+		}
+		if (UB.removeNode(x)) {
+			if (reactOnModification) {
+				delta.add(x, GraphDelta.NR, cause);
+			}
+			GraphEventType e = GraphEventType.REMOVE_NODE;
+			notifyPropagators(e, cause);
+			return true;
+		}
+		return false;
+	}
 
-    /**
-     * Enforce the node x to belong to any solution
-     * Adds x to the lower bound graph
-     *
-     * @param x     node's index
-     * @param cause algorithm which is related to the modification
-     * @return true iff the enforcing has an effect
-     */
-    public boolean enforceNode(int x, ICause cause) throws ContradictionException {
-        assert cause != null;
-        assert (x >= 0 && x < n);
-        if (UB.getNodes().contains(x)) {
-            if (LB.addNode(x)) {
-                if (reactOnModification) {
-                    delta.add(x, GraphDelta.NE, cause);
-                }
-                GraphEventType e = GraphEventType.ADD_NODE;
-                notifyPropagators(e, cause);
-                return true;
-            }
-            return false;
-        }
-        this.contradiction(cause,"enforce node which is not in the domain");
-        return true;
-    }
+	/**
+	 * Enforce the node x to belong to any solution
+	 * Adds x to the lower bound graph
+	 *
+	 * @param x     node's index
+	 * @param cause algorithm which is related to the modification
+	 * @return true iff the enforcing has an effect
+	 */
+	public boolean enforceNode(int x, ICause cause) throws ContradictionException {
+		assert cause != null;
+		assert (x >= 0 && x < n);
+		if (UB.getNodes().contains(x)) {
+			if (LB.addNode(x)) {
+				if (reactOnModification) {
+					delta.add(x, GraphDelta.NE, cause);
+				}
+				GraphEventType e = GraphEventType.ADD_NODE;
+				notifyPropagators(e, cause);
+				return true;
+			}
+			return false;
+		}
+		this.contradiction(cause,"enforce node which is not in the domain");
+		return true;
+	}
 
-    /**
-     * Remove arc (or edge in case of undirected graph variable) (x,y) from the domain
-     * Removes (x,y) from the upper bound graph
-     *
-     * @param x     node's index
-     * @param y     node's index
-     * @param cause algorithm which is related to the removal
-     * @return true iff the removal has an effect
-     * @throws ContradictionException if the arc was mandatory
-     */
-    public abstract boolean removeArc(int x, int y, ICause cause) throws ContradictionException;
+	/**
+	 * Remove arc (or edge in case of undirected graph variable) (x,y) from the domain
+	 * Removes (x,y) from the upper bound graph
+	 *
+	 * @param x     node's index
+	 * @param y     node's index
+	 * @param cause algorithm which is related to the removal
+	 * @return true iff the removal has an effect
+	 * @throws ContradictionException if the arc was mandatory
+	 */
+	public abstract boolean removeArc(int x, int y, ICause cause) throws ContradictionException;
 
-    /**
-     * Enforces arc (or edge in case of undirected graph variable) (x,y) to belong to any solution
-     * Adds (x,y) to the lower bound graph
-     *
-     * @param x     node's index
-     * @param y     node's index
-     * @param cause algorithm which is related to the removal
-     * @return true iff the enforcing has an effect
-     */
-    public abstract boolean enforceArc(int x, int y, ICause cause) throws ContradictionException;
+	/**
+	 * Enforces arc (or edge in case of undirected graph variable) (x,y) to belong to any solution
+	 * Adds (x,y) to the lower bound graph
+	 *
+	 * @param x     node's index
+	 * @param y     node's index
+	 * @param cause algorithm which is related to the removal
+	 * @return true iff the enforcing has an effect
+	 */
+	public abstract boolean enforceArc(int x, int y, ICause cause) throws ContradictionException;
 
-    //***********************************************************************************
-    // ACCESSORS
-    //***********************************************************************************
+	//***********************************************************************************
+	// ACCESSORS
+	//***********************************************************************************
 
 	/**
 	 * @return the lower bound graph (having mandatory nodes and arcs)
 	 */
-    public E getLB() {
-        return LB;
-    }
+	public E getLB() {
+		return LB;
+	}
 
 	/**
 	 * @return the upper bound graph (having possible nodes and arcs)
 	 */
-    public E getUB() {
-        return UB;
-    }
+	public E getUB() {
+		return UB;
+	}
 
-    /**
-     * Get the set of successors (if directed) or neighbors (if undirected) of vertex 'idx'
-     * in the lower bound graph (mandatory outgoing arcs)
-     * @param idx	a vertex
-     * @return The set of successors (if directed) or neighbors (if undirected) of 'idx' in LB
-     */
-    public ISet getMandSuccOrNeighOf(int idx) {
-        return LB.getSuccOrNeighOf(idx);
-    }
+	/**
+	 * Get the set of successors (if directed) or neighbors (if undirected) of vertex 'idx'
+	 * in the lower bound graph (mandatory outgoing arcs)
+	 * @param idx	a vertex
+	 * @return The set of successors (if directed) or neighbors (if undirected) of 'idx' in LB
+	 */
+	public ISet getMandSuccOrNeighOf(int idx) {
+		return LB.getSuccOrNeighOf(idx);
+	}
 
-    /**
-     * Get the set of successors (if directed) or neighbors (if undirected) of vertex 'idx'
-     * in the upper bound graph (potential outgoing arcs)
-     * @param idx	a vertex
-     * @return The set of successors (if directed) or neighbors (if undirected) of 'idx' in UB
-     */
-    public ISet getPotSuccOrNeighOf(int idx) {
-        return UB.getSuccOrNeighOf(idx);
-    }
+	/**
+	 * Get the set of successors (if directed) or neighbors (if undirected) of vertex 'idx'
+	 * in the upper bound graph (potential outgoing arcs)
+	 * @param idx	a vertex
+	 * @return The set of successors (if directed) or neighbors (if undirected) of 'idx' in UB
+	 */
+	public ISet getPotSuccOrNeighOf(int idx) {
+		return UB.getSuccOrNeighOf(idx);
+	}
 
-    /**
-     * Get the set of predecessors (if directed) or neighbors (if undirected) of vertex 'idx'
-     * in the lower bound graph (mandatory ingoing arcs)
-     * @param idx	a vertex
-     * @return The set of predecessors (if directed) or neighbors (if undirected) of 'idx' in LB
-     */
-    public ISet getMandPredOrNeighOf(int idx) {
-        return LB.getPredOrNeighOf(idx);
-    }
+	/**
+	 * Get the set of predecessors (if directed) or neighbors (if undirected) of vertex 'idx'
+	 * in the lower bound graph (mandatory ingoing arcs)
+	 * @param idx	a vertex
+	 * @return The set of predecessors (if directed) or neighbors (if undirected) of 'idx' in LB
+	 */
+	public ISet getMandPredOrNeighOf(int idx) {
+		return LB.getPredOrNeighOf(idx);
+	}
 
-    /**
-     * Get the set of predecessors (if directed) or neighbors (if undirected) of vertex 'idx'
-     * in the upper bound graph (potential ingoing arcs)
-     * @param idx	a vertex
-     * @return The set of predecessors (if directed) or neighbors (if undirected) of 'idx' in UB
-     */
-    public ISet getPotPredOrNeighOf(int idx) {
-        return UB.getPredOrNeighOf(idx);
-    }
+	/**
+	 * Get the set of predecessors (if directed) or neighbors (if undirected) of vertex 'idx'
+	 * in the upper bound graph (potential ingoing arcs)
+	 * @param idx	a vertex
+	 * @return The set of predecessors (if directed) or neighbors (if undirected) of 'idx' in UB
+	 */
+	public ISet getPotPredOrNeighOf(int idx) {
+		return UB.getPredOrNeighOf(idx);
+	}
 
-    /**
-     * @return the maximum number of node the graph variable may have.
-     * Nodes are comprised in the interval [0,getNbMaxNodes()]
-     * Therefore, any vertex should be strictly lower than getNbMaxNodes()
-     */
-    public int getNbMaxNodes() {
-        return n;
-    }
+	/**
+	 * @return the maximum number of node the graph variable may have.
+	 * Nodes are comprised in the interval [0,getNbMaxNodes()]
+	 * Therefore, any vertex should be strictly lower than getNbMaxNodes()
+	 */
+	public int getNbMaxNodes() {
+		return n;
+	}
 
-    /**
-     * @return the node set of the lower bound graph,
-     * i.e. nodes that belong to every solution
-     */
-    public ISet getMandatoryNodes() {
-        return LB.getNodes();
-    }
+	/**
+	 * @return the node set of the lower bound graph,
+	 * i.e. nodes that belong to every solution
+	 */
+	public ISet getMandatoryNodes() {
+		return LB.getNodes();
+	}
 
-    /**
-     * @return the node set of the upper bound graph,
-     * i.e. nodes that may belong to one solution
-     */
-    public ISet getPotentialNodes() {
-        return UB.getNodes();
-    }
+	/**
+	 * @return the node set of the upper bound graph,
+	 * i.e. nodes that may belong to one solution
+	 */
+	public ISet getPotentialNodes() {
+		return UB.getNodes();
+	}
 
 	/**
 	 * @return true iff the graph is directed. It is undirected otherwise.
 	 */
-    public abstract boolean isDirected();
+	public abstract boolean isDirected();
 
-    //***********************************************************************************
-    // {T, U, F} GRAPH VAR REPRESENTATION
-    // cf. Beldiceanu et al. 2006 (https://link.springer.com/chapter/10.1007/11889205_7)
-    //***********************************************************************************
+	//***********************************************************************************
+	// {T, U, F} GRAPH VAR REPRESENTATION
+	// cf. Beldiceanu et al. 2006 (https://link.springer.com/chapter/10.1007/11889205_7)
+	//***********************************************************************************
 
-    public enum NodeSet {
-        T, U, TU
-    }
+	public enum NodeSet {
+		T, U, TU
+	}
 
-    /**
-     * @return The T-nodes of the graph variable.
-     */
-    public ISet getNodes(NodeSet nodeSet) {
-        switch (nodeSet) {
-            case T:
-                return getTNodes();
-            case TU:
-                return getTUNodes();
-            case U:
-                return getUNodes();
-            default:
-                return getTUNodes();
-        }
-    }
+	/**
+	 * @return The T-nodes of the graph variable.
+	 */
+	public ISet getNodes(NodeSet nodeSet) {
+		switch (nodeSet) {
+			case T:
+				return getTNodes();
+			case TU:
+				return getTUNodes();
+			case U:
+				return getUNodes();
+			default:
+				return getTUNodes();
+		}
+	}
 
-    protected ISet getTNodes() {
-        return LB.getNodes();
-    }
+	protected ISet getTNodes() {
+		return LB.getNodes();
+	}
 
-    protected ISet getUNodes() {
-        ISet tNodes = getTNodes();
-        ISet tuNodes = getTUNodes();
-        ISet uNodes = SetFactory.makeBitSet(0);
-        for (int e : tuNodes) {
-            if (!tNodes.contains(e)) {
-                uNodes.add(e);
-            }
-        }
-        return uNodes;
-    }
+	protected ISet getUNodes() {
+		ISet tNodes = getTNodes();
+		ISet tuNodes = getTUNodes();
+		ISet uNodes = SetFactory.makeBitSet(0);
+		for (int e : tuNodes) {
+			if (!tNodes.contains(e)) {
+				uNodes.add(e);
+			}
+		}
+		return uNodes;
+	}
 
-    protected ISet getTUNodes() {
-        return UB.getNodes();
-    }
+	protected ISet getTUNodes() {
+		return UB.getNodes();
+	}
 
-    //***********************************************************************************
-    // VARIABLE STUFF
-    //***********************************************************************************
+	//***********************************************************************************
+	// VARIABLE STUFF
+	//***********************************************************************************
 
 
-    @Override
-    public GraphDelta getDelta() {
-        return delta;
-    }
+	@Override
+	public GraphDelta getDelta() {
+		return delta;
+	}
 
-    @Override
-    public int getTypeAndKind() {
-        return VAR | GRAPH;
-    }
+	@Override
+	public int getTypeAndKind() {
+		return VAR | GRAPH;
+	}
 
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("graph_var ").append(getName());
-        if (isInstantiated()) {
-            sb.append("\nValue: \n");
-            sb.append(UB.toString());
-        } else {
-            sb.append("\nUpper bound: \n");
-            sb.append(UB.toString());
-            sb.append("\nLower bound: \n");
-            sb.append(LB.toString());
-        }
-        return sb.toString();
-    }
+	@Override
+	public String toString() {
+		StringBuilder sb = new StringBuilder();
+		sb.append("graph_var ").append(getName());
+		if (isInstantiated()) {
+			sb.append("\nValue: \n");
+			sb.append(UB.toString());
+		} else {
+			sb.append("\nUpper bound: \n");
+			sb.append(UB.toString());
+			sb.append("\nLower bound: \n");
+			sb.append(LB.toString());
+		}
+		return sb.toString();
+	}
 
-    @Override
-    public void createDelta() {
-        if (!reactOnModification) {
-            reactOnModification = true;
-            delta = new GraphDelta(getEnvironment());
-        }
-    }
+	@Override
+	public void createDelta() {
+		if (!reactOnModification) {
+			reactOnModification = true;
+			delta = new GraphDelta(getEnvironment());
+		}
+	}
 
-    /**
-     * Make the propagator 'prop' have an incremental filtering w.r.t. this graph variable
-     * @param propagator A propagator involving this graph variable
-     * @return A new instance of GraphDeltaMonitor to make incremental propagators
-     */
-    public GraphDeltaMonitor monitorDelta(ICause propagator) {
-        createDelta();
-        return new GraphDeltaMonitor(delta, propagator);
-    }
+	/**
+	 * Make the propagator 'prop' have an incremental filtering w.r.t. this graph variable
+	 * @param propagator A propagator involving this graph variable
+	 * @return A new instance of GraphDeltaMonitor to make incremental propagators
+	 */
+	public GraphDeltaMonitor monitorDelta(ICause propagator) {
+		createDelta();
+		return new GraphDeltaMonitor(delta, propagator);
+	}
 
-    @Override
-    public void notifyMonitors(IEventType event) throws ContradictionException {
-        for (int i = mIdx - 1; i >= 0; i--) {
-            monitors[i].onUpdate(this, event);
-        }
-    }
+	@Override
+	public void notifyMonitors(IEventType event) throws ContradictionException {
+		for (int i = mIdx - 1; i >= 0; i--) {
+			monitors[i].onUpdate(this, event);
+		}
+	}
 
-    //***********************************************************************************
-    // SOLUTIONS : STORE AND RESTORE
-    //***********************************************************************************
+	//***********************************************************************************
+	// SOLUTIONS : STORE AND RESTORE
+	//***********************************************************************************
 
 	/**
 	 * @return the value of the graph variable represented through an adjacency matrix
 	 *         plus a set of nodes (last row of the matrix).
 	 *         This method is not supposed to be used except for restoring solutions.
 	 */
-    public boolean[][] getValue() {
-        int n = getUB().getNbMaxNodes();
-        boolean[][] vals = new boolean[n + 1][n];
-        for (int i : getLB().getNodes()) {
-            for (int j : getLB().getSuccOrNeighOf(i)) {
-                vals[i][j] = true; // arc in
-            }
-            vals[n][i] = true; // node in
-        }
-        return vals;
-    }
+	public boolean[][] getValue() {
+		int n = getUB().getNbMaxNodes();
+		boolean[][] vals = new boolean[n + 1][n];
+		for (int i : getLB().getNodes()) {
+			for (int j : getLB().getSuccOrNeighOf(i)) {
+				vals[i][j] = true; // arc in
+			}
+			vals[n][i] = true; // node in
+		}
+		return vals;
+	}
 
 	/**
 	 * Instantiates <code>this</code> to value which represents an adjacency
@@ -405,23 +405,23 @@ public abstract class GraphVar<E extends IGraph> extends AbstractVariable implem
 	 * @param cause
 	 * @throws ContradictionException if the arc was mandatory
 	 */
-    public void instantiateTo(boolean[][] value, ICause cause) throws ContradictionException {
-        int n = value.length - 1;
-        for (int i = 0; i < n; i++) {
-            if (value[n][i]) {//nodes
-                enforceNode(i, cause);
-            } else {
-                removeNode(i, cause);
-            }
-            for (int j = 0; j < n; j++) {
-                if (value[i][j]) {//arcs
-                    enforceArc(i, j, cause);
-                } else {
-                    removeArc(i, j, cause);
-                }
-            }
-        }
-    }
+	public void instantiateTo(boolean[][] value, ICause cause) throws ContradictionException {
+		int n = value.length - 1;
+		for (int i = 0; i < n; i++) {
+			if (value[n][i]) {//nodes
+				enforceNode(i, cause);
+			} else {
+				removeNode(i, cause);
+			}
+			for (int j = 0; j < n; j++) {
+				if (value[i][j]) {//arcs
+					enforceArc(i, j, cause);
+				} else {
+					removeArc(i, j, cause);
+				}
+			}
+		}
+	}
 
 	//***********************************************************************************
 	// GraphViz
